move ContrastInfoCard css into ContrastInfoCard.module.css
